fix(trace): avoid report audit drift
