+++ src/core/extension/trace/mod.rs
@@ -17,6 +17,7 @@ pub mod parsing;
 pub mod probes;
 pub mod report;
 pub mod run;
+mod span_summary;
 pub mod spans;
 
 use crate::core::component::Component;
@@ -35,19 +36,19 @@ pub use parsing::{
 pub use parsing::{TraceAssertionStatus, TraceResults, TraceSpanDefinition, TraceSpanResult};
 pub use probes::{ActiveTraceProbes, TraceProbeConfig};
 pub use report::{
-    attach_span_summary_metadata, from_list_workflow, from_main_workflow,
-    from_main_workflow_outputs, TraceAggregateOutput, TraceAggregateRunOutput,
-    TraceAggregateSpanOutput, TraceClassificationSummaryOutput, TraceCommandOutput,
-    TraceCompareClassificationSummaryOutput, TraceCompareOutput, TraceCompareSpanOutput,
-    TraceGuardrailOutput, TraceListOutput, TraceOverlayLocksOutput, TraceProfileListItem,
-    TraceResolvedProfileOutput, TraceRunOrderEntryOutput, TraceSpanMetadata,
-    TraceSpanSummaryOutput, TraceVariantMatrixOutput, TraceVariantMatrixRunOutput,
+    from_list_workflow, from_main_workflow, from_main_workflow_outputs, TraceAggregateOutput,
+    TraceAggregateRunOutput, TraceAggregateSpanOutput, TraceClassificationSummaryOutput,
+    TraceCommandOutput, TraceCompareClassificationSummaryOutput, TraceCompareOutput,
+    TraceCompareSpanOutput, TraceGuardrailOutput, TraceListOutput, TraceOverlayLocksOutput,
+    TraceProfileListItem, TraceResolvedProfileOutput, TraceRunOrderEntryOutput, TraceSpanMetadata,
+    TraceVariantMatrixOutput, TraceVariantMatrixRunOutput,
 };
 pub use report::{push_overlay_markdown, render_markdown};
 pub use run::{
     run_trace_list_workflow, run_trace_workflow, trace_is_unclaimed, TraceListWorkflowArgs,
 };
 pub use run::{TraceRunWorkflowArgs, TraceRunWorkflowResult, TraceRunnerInputs};
+pub use span_summary::{attach_span_summary_metadata, TraceSpanSummaryOutput};
 
 pub fn resolve_trace_command(
     component: &Component,
